chore: remove Python 2 compatibility from tests
