+++ setup.cfg
@@ -25,9 +25,9 @@ keywords =
 [options]
 zip_safe = True
 install_requires =
-    importlib-resources
     click
     click-log
+    importlib-resources
     ply
     SQLAlchemy
     tqdm
