+++ tests/test_integration/test_brenda_parser.py
@@ -1,4 +1,3 @@
-
 # from click.testing import CliRunner
 
 # from brenda_parser.cli import main
